DefaultValue is checked for conflicts (#44)
Previously, `DefaultValue` was a No-op that could be set multiple times.
This could lead to unexpected results when we request a default value,
but accidentally overwrite it with a concrete value.

This PR changes the behavior, so that placeholders are kept until the
call to `initialize` and checked for conflicts. That way, the behavior
should be more consistent overall.